chore: rename deployment env helper
Rename the environment detection helper to a more generic deploymentEnv module.

- move vercelEnv to deploymentEnv
- update imports to use the new module name
