+++ middleware.ts
@@ -3,7 +3,7 @@
 import { createServerClient } from '@supabase/ssr';
 import { NextResponse, type NextRequest } from 'next/server';
 import { getStoreConfig } from './src/server/storeConfig';
-import { isPreviewDeployment } from './src/server/vercelEnv';
+import { isPreviewDeployment } from './src/server/deploymentEnv';
 
 // Fail fast if the deployment hasn't selected a provenance store.
 getStoreConfig();
