+++ src/server/auth.ts
@@ -5,7 +5,7 @@ import { unauthorized } from '@/src/server/http';
 import { createSupabaseServerClient } from '@/src/server/supabase/server';
 import { assertLocalPgModeConfig, isLocalPgMode } from '@/src/server/pgMode';
 import { LOCAL_PG_USER_ID } from '@/src/server/localPgConfig';
-import { isPreviewDeployment } from '@/src/server/vercelEnv';
+import { isPreviewDeployment } from '@/src/server/deploymentEnv';
 
 const LOCAL_USER: User = {
   id: LOCAL_PG_USER_ID,
